+++ cloudant-sync-datastore-core/src/test/java/com/cloudant/sync/replication/PushReplicatorTest.java
@@ -336,4 +336,29 @@ public class PushReplicatorTest extends ReplicationTestBase {
         assertLastSequence(replicator);
     }
 
+    /**
+     * Test that a username and password combination where both parts contain a series of URI
+     * reserved and other percent encoded characters is correctly encoded and not double encoded
+     * after going through the ReplicatorBuilder and CookieInterceptor.
+     *
+     * @throws Exception
+     */
+    @Test
+    public void replicatorBuilderAddsCookieInterceptorCredsPercentEncoded() throws Exception {
+        String encodedUsername = "user" + PERCENT_ENCODED_URI_CHARS;
+        String encodedPassword = "password" + PERCENT_ENCODED_URI_CHARS;
+        ReplicatorBuilder.Push p = ReplicatorBuilder.push().
+                from(datastore).
+                to(new URI("http://" + encodedUsername + ":" + encodedPassword +
+                        "@some-host/path%2Fsome-path-日本"));
+        ReplicatorImpl r = (ReplicatorImpl) p.build();
+        // check that user/pass has been removed
+        Assert.assertEquals("http://some-host:80/path%2Fsome-path-日本",
+                (((CouchClientWrapper) (((PushStrategy) r.strategy).targetDb)).
+                        getCouchClient().
+                        getRootUri()).
+                        toString()
+        );
+        assertCookieInterceptorPresent(p, "name="+encodedUsername+"&password=" + encodedPassword);
+    }
 }