+++ CHANGES.md
@@ -1,3 +1,7 @@
+# Unreleased
+ - [FIXED] Issue with double encoding of restricted URL characters in credentials when using
+   `ReplicatorBuilder`.
+
 # 1.1.5 (2016-12-08)
 - [FIXED] Issue where replicator would not get the latest revision if `_bulk_get`
    was available.